feature: create digiwf-unit-test lib and rename digiwf-integration-e2e-test to digiwf-e2e-test
