+++ digiwf-integrations/digiwf-cosys-integration/digiwf-cosys-integration-client/pom.xml
@@ -48,7 +48,7 @@
         <!-- testing libraries-->
         <dependency>
             <groupId>de.muenchen.oss.digiwf</groupId>
-            <artifactId>digiwf-integration-e2e-test-starter</artifactId>
+            <artifactId>digiwf-e2e-test-starter</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
