+++ digiwf-integrations/digiwf-dms-integration/digiwf-dms-integration-core/pom.xml
@@ -74,7 +74,7 @@
         </dependency>
         <dependency>
             <groupId>de.muenchen.oss.digiwf</groupId>
-            <artifactId>digiwf-integration-e2e-test-starter</artifactId>
+            <artifactId>digiwf-e2e-test-starter</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
