+++ digiwf-integrations/digiwf-dms-integration/digiwf-dms-integration-fabasoft/digiwf-dms-integration-fabasoft-mock/pom.xml
@@ -19,7 +19,7 @@
         </dependency>
         <dependency>
             <groupId>de.muenchen.oss.digiwf</groupId>
-            <artifactId>digiwf-integration-e2e-test-starter</artifactId>
+            <artifactId>digiwf-e2e-test-starter</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
