--- digiwf-integrations/digiwf-okewo-integration/digiwf-okewo-integration-service/src/test/java/de/muenchen/oss/digiwf/okewo/integration/EmbeddedKafkaConfig.java
@@ -1,24 +1,0 @@
-package de.muenchen.oss.digiwf.okewo.integration;
-
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.test.EmbeddedKafkaBroker;
-
-@Configuration
-public class EmbeddedKafkaConfig {
-
-    @Bean
-    @ConditionalOnMissingBean
-    public EmbeddedKafkaBroker embeddedKafkaConfig() {
-        return new EmbeddedKafkaBroker(
-                1,
-                true,
-                "dwf-okewo-e2e-test",
-                "dwf-connector-e2e-test",
-                "dwf-connector-bpmnerror-e2e-test",
-                "dwf-connector-incident-e2e-test"
-        );
-    }
-
-}
