+++ digiwf-integrations/digiwf-okewo-integration/digiwf-okewo-integration-service/src/test/java/de/muenchen/oss/digiwf/okewo/integration/OkEwoIntegrationApplicationE2eTest.java
@@ -2,47 +2,36 @@ package de.muenchen.oss.digiwf.okewo.integration;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.junit5.WireMockTest;
+import de.muenchen.oss.digiwf.integration.e2e.test.DigiwfE2eTest;
+import de.muenchen.oss.digiwf.integration.e2e.test.DigiwfIntegrationE2eTestUtility;
+import de.muenchen.oss.digiwf.integration.e2e.test.DigiwfWiremockUtility;
 import de.muenchen.oss.digiwf.okewo.integration.client.model.*;
 import de.muenchen.oss.digiwf.okewo.integration.domain.model.request.OkEwoOmBasedRequest;
 import de.muenchen.oss.digiwf.okewo.integration.domain.model.request.OkEwoSearchPersonExtendedRequest;
 import de.muenchen.oss.digiwf.okewo.integration.domain.model.request.OkEwoSearchPersonRequest;
 import de.muenchen.oss.digiwf.okewo.integration.domain.model.request.OrdnungsmerkmalDto;
-import de.muenchen.oss.digiwf.okewo.integration.utility.DigiWFIntegrationE2eTest;
 import lombok.val;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.kafka.test.context.EmbeddedKafka;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ActiveProfiles;
 
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.equalToJson;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("itest")
-@DirtiesContext
-@EmbeddedKafka(partitions = 1,
-    brokerProperties = {"listeners=PLAINTEXT://localhost:29092"},
-    topics = {
-        "${spring.cloud.stream.bindings.functionRouter-in-0.destination}",
-        "${spring.cloud.stream.bindings.sendMessage-out-0.destination}",
-        "${spring.cloud.stream.bindings.integrationTestConsumer-in-0.destination}"
-    })
+@DigiwfE2eTest
 @WireMockTest(httpPort = 8189)
-class OkEwoIntegrationApplicationE2eTest extends DigiWFIntegrationE2eTest {
+class OkEwoIntegrationApplicationE2eTest {
 
 
   @Autowired
   private ObjectMapper objectMapper;
+  @Autowired
+  private DigiwfIntegrationE2eTestUtility digiWFIntegrationE2eTestUtility;
 
   private String processInstanceId;
 
@@ -64,10 +53,10 @@ class OkEwoIntegrationApplicationE2eTest extends DigiWFIntegrationE2eTest {
     request.setRequest(ordnungsmerkmal);
 
     val person = new Person().ordnungsmerkmal("om");
-    this.setupGetPathOfWiremock("/personen/2.0/rest/person/om?benutzerId=benutzerId", "username", "password", objectMapper.writeValueAsString(person));
+    DigiwfWiremockUtility.setupGETWithBasicAuth("/personen/2.0/rest/person/om?benutzerId=benutzerId", "username", "password", objectMapper.writeValueAsString(person));
 
     // send and receive messages
-    final Map<String, Object> payload = super.runIntegration(request, processInstanceId, "getPerson");
+    final Map<String, Object> payload = digiWFIntegrationE2eTestUtility.runIntegration(request, processInstanceId, "getPerson");
     assertNotNull(payload);
     final Map<String, Object> response = (Map<String, Object>) payload.get("response");
     assertNotNull(response);
@@ -86,10 +75,10 @@ class OkEwoIntegrationApplicationE2eTest extends DigiWFIntegrationE2eTest {
     val person = new Person().ordnungsmerkmal("om");
     val apiResponse = new SuchePersonAntwort().personen(List.of(person));
 
-    this.setupPostPathOfWiremock("/personen/2.0/rest/person/search", objectMapper.writeValueAsString(searchRequest), "username", "password", objectMapper.writeValueAsString(apiResponse));
+    DigiwfWiremockUtility.setupPOSTWithBasicAuth("/personen/2.0/rest/person/search", objectMapper.writeValueAsString(searchRequest), "username", "password", objectMapper.writeValueAsString(apiResponse));
 
     // send and receive messages
-    final Map<String, Object> payload = super.runIntegration(request, processInstanceId, "searchPerson");
+    final Map<String, Object> payload = digiWFIntegrationE2eTestUtility.runIntegration(request, processInstanceId, "searchPerson");
     assertNotNull(payload);
     final Map<String, Object> response = (Map<String, Object>) payload.get("response");
     assertNotNull(response);
@@ -105,10 +94,10 @@ class OkEwoIntegrationApplicationE2eTest extends DigiWFIntegrationE2eTest {
     request.setRequest(ordnungsmerkmal);
 
     val person = new PersonErweitert().ordnungsmerkmal("om");
-    this.setupGetPathOfWiremock("/personen/2.0/rest/personErweitert/om?benutzerId=benutzerId", "username", "password", objectMapper.writeValueAsString(person));
+    DigiwfWiremockUtility.setupGETWithBasicAuth("/personen/2.0/rest/personErweitert/om?benutzerId=benutzerId", "username", "password", objectMapper.writeValueAsString(person));
 
     // send and receive messages
-    final Map<String, Object> payload = super.runIntegration(request, processInstanceId, "getPersonErweitert");
+    final Map<String, Object> payload = digiWFIntegrationE2eTestUtility.runIntegration(request, processInstanceId, "getPersonErweitert");
     assertNotNull(payload);
     final Map<String, Object> response = (Map<String, Object>) payload.get("response");
     assertNotNull(response);
@@ -127,10 +116,10 @@ class OkEwoIntegrationApplicationE2eTest extends DigiWFIntegrationE2eTest {
     val person = new PersonErweitert().ordnungsmerkmal("om");
     val apiResponse = new SuchePersonerweitertAntwort().personen(List.of(person));
 
-    this.setupPostPathOfWiremock("/personen/2.0/rest/personErweitert/search", objectMapper.writeValueAsString(searchRequest), "username", "password", objectMapper.writeValueAsString(apiResponse));
+    DigiwfWiremockUtility.setupPOSTWithBasicAuth("/personen/2.0/rest/personErweitert/search", objectMapper.writeValueAsString(searchRequest), "username", "password", objectMapper.writeValueAsString(apiResponse));
 
     // send and receive messages
-    final Map<String, Object> payload = super.runIntegration(request, processInstanceId, "searchPersonErweitert");
+    final Map<String, Object> payload = digiWFIntegrationE2eTestUtility.runIntegration(request, processInstanceId, "searchPersonErweitert");
     assertNotNull(payload);
     final Map<String, Object> response = (Map<String, Object>) payload.get("response");
     assertNotNull(response);
@@ -138,26 +127,4 @@ class OkEwoIntegrationApplicationE2eTest extends DigiWFIntegrationE2eTest {
     assertEquals(1, personenResponse.size());
   }
 
-  private void setupGetPathOfWiremock(final String url, final String username, final String password, final String expectedResponse) {
-    WireMock.stubFor(WireMock
-        .get(url)
-        .withBasicAuth(username, password)
-        .willReturn(WireMock
-            .aResponse()
-            .withBody(expectedResponse)
-            .withHeader("Content-Type", "application/json")
-            .withStatus(200)));
-  }
-
-  private void setupPostPathOfWiremock(final String url, final String requestBody, final String username, final String password, final String expectedResponse) {
-    WireMock.stubFor(WireMock
-        .post(url)
-        .withRequestBody(equalToJson(requestBody))
-        .withBasicAuth(username, password)
-        .willReturn(WireMock
-            .aResponse()
-            .withBody(expectedResponse)
-            .withHeader("Content-Type", "application/json")
-            .withStatus(200)));
-  }
 }
