+++ digiwf-libs/digiwf-testing/REAMDE.md
@@ -9,7 +9,7 @@ Add the following dependency to your project:
 ```xml
         <dependency>
             <groupId>de.muenchen.oss.digiwf</groupId>
-            <artifactId>digiwf-integration-e2e-test-starter</artifactId>
+            <artifactId>digiwf-e2e-test-starter</artifactId>
             <version>${latest.digiwf.version}</version>
             <scope>test</scope>
         </dependency>
