+++ digiwf-libs/digiwf-testing/digiwf-e2e-test-core/pom.xml
@@ -5,11 +5,11 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>de.muenchen.oss.digiwf</groupId>
-        <artifactId>digiwf-integration-e2e-test</artifactId>
+        <artifactId>digiwf-testing</artifactId>
         <version>1.4.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>digiwf-integration-e2e-test-core</artifactId>
+    <artifactId>digiwf-e2e-test-core</artifactId>
 
     <dependencies>
         <dependency>
