+++ digiwf-libs/digiwf-testing/digiwf-e2e-test-core/src/main/java/de/muenchen/oss/digiwf/integration/e2e/test/wsdl/DigiwfWiremockWsdlUtility.java
