+++ digiwf-libs/digiwf-testing/digiwf-e2e-test-starter/src/main/java/de/muenchen/oss/digiwf/integration/e2e/test/autoconfiguration/E2eAutoConfiguration.java
