+++ digiwf-libs/digiwf-testing/digiwf-e2e-test-starter/src/main/resources/integration-e2e-test-application.yaml
