+++ digiwf-libs/pom.xml
@@ -24,7 +24,7 @@
         <module>digiwf-spring-security</module>
         <module>digiwf-spring-logging-and-tracing</module>
         <module>digiwf-email</module>
-        <module>digiwf-integration-e2e-test</module>
+        <module>digiwf-testing</module>
         <module>digiwf-process</module>
     </modules>
